:fire: Remove animation related code in OptionDrawer
