+++ src/molecules/OptionDrawer/OptionDrawer.test.tsx
@@ -5,12 +5,7 @@ import {
   OptionDrawerProps,
   ToggleEventProps,
 } from 'src/molecules/OptionDrawer/OptionDrawer';
-import {
-  render,
-  screen,
-  userEvent,
-  waitFor,
-} from 'src/tests/browsertest-utils';
+import { render, screen, userEvent } from 'src/tests/browsertest-utils';
 
 interface FakeType {
   id: string;
@@ -164,185 +159,6 @@ test('Shows expected status for parent if singleSelect=true', async () => {
     'data-indeterminate',
     'false'
   );
-});
-
-test('Animation works correctly when toggling parent with no child', async () => {
-  const props = fakeProps(false, false);
-  const user = userEvent.setup();
-  const { rerender } = render(<OptionDrawer {...props} animateCheck={true} />);
-
-  const parentElement = screen.getByText(props.item.label);
-
-  await user.click(parentElement);
-
-  await waitFor(
-    () =>
-      expect(
-        screen.getByTestId('animated-' + props.item.id).parentElement
-      ).toBeInTheDocument(),
-    { timeout: 500 }
-  );
-
-  rerender(
-    <OptionDrawer
-      {...props}
-      selectedItems={[props.item]}
-      animateUncheck={true}
-    />
-  );
-
-  await user.click(parentElement);
-
-  await waitFor(
-    () =>
-      expect(
-        screen.getByTestId('animated-' + props.item.id).parentElement
-      ).toBeInTheDocument(),
-    { timeout: 500 }
-  );
-});
-
-test('Children does not trigger animation if you check all but one sibling', async () => {
-  const props = fakeProps();
-  const user = userEvent.setup();
-  const children = props.item.children ?? [props.item];
-  const childrenExpectLast = children.slice(0, -1);
-
-  render(<OptionDrawer {...props} animateCheck={true} openAll />);
-
-  for (const child of childrenExpectLast) {
-    await user.click(screen.getByText(child.label));
-  }
-
-  expect(screen.queryByTestId('animated-' + props.item.id)).toBeNull();
-  expect(props.onToggle).toHaveBeenCalledTimes(childrenExpectLast.length);
-});
-
-test('Children does not trigger animation if you uncheck all but one sibling', async () => {
-  const props = fakeProps();
-  const user = userEvent.setup();
-  const children = props.item.children ?? [props.item];
-  const childrenExpectLast = children.slice(0, -1);
-
-  render(
-    <OptionDrawer
-      {...props}
-      selectedItems={children}
-      animateUncheck={true}
-      openAll
-    />
-  );
-
-  for (const child of childrenExpectLast) {
-    await user.click(screen.getByText(child.label));
-  }
-
-  expect(screen.queryByTestId('animated-' + props.item.id)).toBeNull();
-  expect(props.onToggle).toHaveBeenCalledTimes(childrenExpectLast.length);
-});
-
-test('Animation works correctly when checking all children', async () => {
-  const props = fakeProps();
-  const user = userEvent.setup();
-  const children = props.item.children ?? [props.item];
-
-  const { rerender } = render(
-    <OptionDrawer
-      {...props}
-      selectedItems={[
-        ...children.filter(
-          (child) => child.id !== children[children.length - 1].id
-        ),
-      ]}
-      animateCheck={true}
-      openAll
-    />
-  );
-
-  await user.click(screen.getByText(children[children.length - 1].label));
-
-  await waitFor(
-    () =>
-      expect(
-        screen.getByTestId('animated-' + props.item.id).parentElement
-      ).toBeInTheDocument(),
-    { timeout: 500 }
-  );
-
-  rerender(
-    <OptionDrawer
-      {...props}
-      selectedItems={[...children]}
-      animateCheck={true}
-      openAll
-    />
-  );
-
-  expect(
-    screen.getByText(props.item.label).children[0].children[0]
-  ).toBeChecked();
-});
-
-test('Animation works correctly when unchecking all children', async () => {
-  const props = fakeProps();
-  const user = userEvent.setup();
-  const children = props.item.children ?? [props.item];
-
-  const { rerender } = render(
-    <OptionDrawer
-      {...props}
-      selectedItems={[children[children.length - 1]]}
-      animateUncheck={true}
-      openAll
-    />
-  );
-  await user.click(screen.getByText(children[children.length - 1].label));
-
-  await waitFor(
-    () =>
-      expect(
-        screen.getByTestId('animated-' + props.item.id).parentElement
-      ).toBeInTheDocument(),
-    { timeout: 500 }
-  );
-
-  rerender(
-    <OptionDrawer {...props} selectedItems={[]} animateCheck={true} openAll />
-  );
-
-  expect(
-    screen.getByText(props.item.label).children[0].children[0]
-  ).not.toBeChecked();
-});
-
-test('onToggle is called after check animation', async () => {
-  const props = fakeProps(false, false);
-  const user = userEvent.setup();
-
-  render(<OptionDrawer {...props} animateCheck={true} />);
-
-  await user.click(screen.getByText(props.item.label));
-  await waitFor(() => expect(props.onToggle).toHaveBeenCalledTimes(1), {
-    timeout: 800,
-  });
-});
-
-test('onToggle is called after uncheck animation', async () => {
-  const props = fakeProps(false, false);
-  const user = userEvent.setup();
-
-  render(
-    <OptionDrawer
-      {...props}
-      selectedItems={[props.item]}
-      animateUncheck={true}
-    />
-  );
-
-  await user.click(screen.getByText(props.item.label));
-  await waitFor(() => expect(props.onToggle).toHaveBeenCalledTimes(1), {
-    timeout: 800,
-  });
 });
 
 test('chevron toggler works as expected', async () => {
